check until output

+++ index.cjs
@@ -1,6 +1,6 @@
-import fs from "fs";
-import path from "path";
-import { translationsFolderName, targetFolderPath, outputFileName } from "./config.js";
+const fs = require("fs");
+const path = require("path");
+const { translationsFolderName, targetFolderPath, outputFileName } = require('./config');
 
 /**
  * Extracts translation keys from the specified file and adds them to the translation object.
@@ -43,6 +43,11 @@ const processFolder = (folderPath, translations) => {
  */
 const main = () => {
     const translations = {};
+    const outputFilePath = path.join(
+        targetFolderPath,
+        translationsFolderName,
+        outputFileName
+    );
 
     // Check the target folder
     if (!fs.existsSync(targetFolderPath)) {
@@ -52,16 +57,29 @@ const main = () => {
         return;
     }
 
+    // Check output file name
+    if (!outputFileName.endsWith(".json")) {
+        console.error(
+            'The output file name must end with ".json". Please specify the correct name in the "config.js" file.'
+        );
+        return;
+    }
+
+    // Check output file path
+    if (fs.existsSync(outputFilePath)) {
+        console.error(
+            `The output file already exists. Please delete it before running the script again: ${outputFilePath}`
+        );
+        return;
+    }
+
     // Recursively process files in the target folder.
     processFolder(targetFolderPath, translations);
 
-    const outputFilePath = path.join(
-        targetFolderPath,
-        translationsFolderName,
-        outputFileName
-    );
+
     const jsonOutput = JSON.stringify(translations, null, 2);
     fs.writeFileSync(outputFilePath, jsonOutput, "utf8");
+    console.log(`Translation keys have been written to ${outputFilePath}.`);
 };
 
 main();
